+++ releasenotes.md
@@ -34,6 +34,7 @@
 * The `r_textures` CVAR no longer affects the menu’s background.
 * A bug is fixed whereby blood splats weren’t as random as intended.
 * When the number of blood splats spawned in the current map reaches the amount set by the `r_bloodsplats_max` CVAR, the oldest blood splats are now removed to make room for new ones, rather than blood splats stopping being spawned altogether.
+* A new read-only CVAR has been implemented called `r_bloodsplats_visible` that shows the number of blood splats currently visible..
 * Monsters that are in liquid sectors now cast shadows when the `r_shadows` CVAR is `on`.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)